app/scripts/requirements.txt
# Dipendenze per gli script Python (preview icone, ecc.)
Pillow>=10.0.0
